yankring_history_dir in the /tmp folder instead of the user folder
Sometime you are editing some files in a terminal session logged as a different user but keeping the environment for the previous user (su -E). Vim will complain that will not be able to write because the yankring_history_dir is read-only (pointing to the previous user). I'd recommend to use the /tmp folder for the yankring_history.